\name{choice.optimal.h}
\docType{methods}
\alias{choice.optimal.h}
\title{Choose the optimal smoothing parameter matrix under the
  cross-validation criteria}
\description{
  This funciton returns the optimal smoothing parameter matrix under the
  cross-validation criteria for Nadaraya-Watson estimator. To minimize
  the cross-validation funciton, this funciton uses the 'optim' funciton
  inside.
}
\section{Methods}{
\describe{
  \item{nw = "nw"}{nw object generated by 'nw' method.}
}}
\keyword{methods}