\name{plot.nw-methods}
\docType{methods}
\alias{plot.nw-methods}
\alias{plot.nw,nw-method}
\title{Plotting the result of regression with Nadaraya-Watson estimator}
\description{
  Plotting the result of regression with Nadaraya-Watson estimator in
  three dimensions.
}
\section{Methods}{
\describe{
  \item{this = "nw"}{nw object generated by 'nw' method.}
}}
\keyword{methods}
\keyword{models}
